ASoC: florida: Add support to sync up to a started ez2control stream
Add support in the trigger callback to check if we have already received
a trigger event before the stream was opened and manually process the
data from this. This will allow streams to be opened in response to the
trigger event if required.

Change-Id: If2ef7631175c5d0245766b5f33a6840cda5ca13b
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
(cherry picked from commit 5c5cd7a84b8a683d64986ff21c49630d34d2bf14)