arm/dt: 8994: bunkerhill p1
Add initial version of device-tree for BunkerHill P1 mortable and
construct the initial Motorola layout for device trees:

  [BunkerHill P1]
  [BunkerHill Common]
  [Moto 8994 Common]

This allows for future products to be added and take advantage of all
the common devices and layout that we'll place in moto common.
BunkerHill common should be empty until we have product differences from
the reference design.

The initial moto-common for 8994 is taken from the QCOM 8994 MTP, up to
commit:

  7b9c8bc power: qpnp-fg: add battery profile loading support

We'll remove and modify as needed to match our actual reference design
from that commit onwards.

Change-Id: Ida588279997dc3b60596b6df7e4f582f39eac75d
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>