msm: mdss: always save display's backlight brightness
Update blank to remember the most recently requested backlight
brightness value, to be re-applied on the first commit after unblank.
This can be either the bl_level or the unset_bl_level, depending on
whether there is an unset_bl_level waiting to be applied.

Waiting until the first commit to turn on the backlight ensures that
the display contents are defined before they are visible.

This handles any of the following sequences correctly:
1. backlight can initially be on or off
2. backlight brightness may be set here
3. blank
4. backlight brightness may be set here
5. unblank
6. backlight brightness may be set here
7. commit, most recent backlight brightness restored from steps 2,4,6

This is a port and update of 91ebbffc.

Change-Id: I832c363eb789234bf80eade37167d825c77c8d43
Signed-off-by: Bang Nguyen <bnguyen1@motorola.com>
Reviewed-on: http://gerrit.mot.com/689618
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>