msm: camera: Add support for actuator LC898212XD
Need to modify the current qc implementation of writing
the i2c registers for the new actuator LC898212xd
since to move the lens requires more i2c registers than
previous simple actuators.

QC will be providing a solution to this with a future
drop, but in the meantime for development this is the
way to move the actuator.

Change-Id: I56c4b588983d60d794a8e96f3a71aff17c541d6d
Signed-off-by: Elena Satraitis <w04888@motorola.com>
Reviewed-on: http://gerrit.mot.com/697377
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Dave Schmidt <d.schmidt@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>