arm/dts: msm8992-moto: Increase Charge Rate
The Charge profile has changed for the battery.

The step voltage needs to be 4.1V and the maximum
current needs to tbe 4.0 A.  Also the Thermal Tables
need to be updated to accomodate the change in rates.

Change-Id: I37d57b4879f379fcfd5c6767c3a7326b2a8274fc
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/742115
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>