arm/dt: 8994: Remove L18 dependency from drv2605
While the L18 does power the mipi_d0 select switch that is used in
powering the vibrator it is also used as a power supply for the Kinzie
display panel. Having the L18 specified as a "static-vdd" to the
vibrator makes it impossible to power down the L18 regulator when the
panel's power supplies need to be cut-off during ESD recovery. By
removing the L18 as a static-vdd supply to the drv2605 vibrator the
driver no longer grabs a reference to the regulator and makes it
possible for the regulator to be powered off when needed.

In practice for Kinzie the L18 regulator is powered for a vast majority
of the time and is only not powered for the extremely rare ESD
recovery sequences of the panel that last for less then 1 second.
This does mean that the vibrator will not function during the ESD
recovery sequence but is an acceptable compromise to properly
follow the panel specifications.

Change-Id: I61e8eb3192b8389f5439e70ba2374293590ff756
Signed-off-by: Joe Swantek <jswantek@motorola.com>
Reviewed-on: http://gerrit.mot.com/750986
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Vladimir Karpovich <vkarpovich@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>